ADEN-8646 Move slot-state-emitter to services
