+++ src/ad-engine/listeners/index.ts
@@ -1,6 +1,5 @@
 export * from './overscroll-listener';
 export * from './porvata-listener';
 export * from './scroll-listener';
-export * from './slot-state-emitter';
 export * from './twitch-listener';
 export * from './listeners';
